chore: disable offset helper
We introduced new functionality & fixed bugs in the OffsetHelper.
These resulted changed ABIs / function names.

The above ends up creating errors when the SDK interacts with
the OffsetHelper. We would need to adapt the SDK to these changes
but right now this is not a priority, so I'm disabling the SDK's
intereactions with it temporarily.